Move structuring code into decompiler/structuring.
