+++ angr/analyses/decompiler/region_simplifiers/cascading_ifs.py
@@ -2,7 +2,8 @@
 import ailment
 
 from ..sequence_walker import SequenceWalker
-from ..structurer_nodes import SequenceNode, CodeNode, MultiNode, LoopNode, ConditionNode, CascadingConditionNode
+from ..structuring.structurer_nodes import SequenceNode, CodeNode, MultiNode, LoopNode, ConditionNode, \
+    CascadingConditionNode
 
 
 class CascadingIfsRemover(SequenceWalker):
