+++ angr/analyses/decompiler/region_simplifiers/if_.py
@@ -4,7 +4,8 @@ import logging
 import ailment
 
 from ..sequence_walker import SequenceWalker
-from ..structurer_nodes import SequenceNode, CodeNode, MultiNode, LoopNode, ConditionNode, CascadingConditionNode
+from ..structuring.structurer_nodes import SequenceNode, CodeNode, MultiNode, LoopNode, ConditionNode, \
+    CascadingConditionNode
 
 
 l = logging.getLogger(name=__name__)
