+++ angr/analyses/decompiler/structuring/structurer_nodes.py
@@ -19,7 +19,7 @@ class MultiNode:
     def __init__(self, nodes):
 
         # delayed import
-        from .graph_region import GraphRegion  # pylint:disable=import-outside-toplevel
+        from ..graph_region import GraphRegion  # pylint:disable=import-outside-toplevel
 
         self.nodes = [ ]
